add tensor times matrix benchmarks to model_trainer. Fix bugs associated with the lapack interface and model trainer

+++ src/shared/model.cxx
@@ -99,7 +99,7 @@ namespace CTF_int {
 //DGELSD computes the minimum-norm solution to a real linear least squares problem:
 //    minimize 2-norm(| b - A*x |)
 //    http://www.netlib.org/lapack/explore-html/d7/d3b/group__double_g_esolve_ga94bd4a63a6dacf523e25ff617719f752.html#ga94bd4a63a6dacf523e25ff617719f752
-  void cdgelsd(int m, int n, int k, double const * A, int lda_A, double * B, int lda_B, double * S, int cond, int * rank, double * work, int lwork, int * iwork, int * info){
+  void cdgelsd(int m, int n, int k, double const * A, int lda_A, double * B, int lda_B, double * S, double cond, int * rank, double * work, int lwork, int * iwork, int * info){
 #ifdef TUNE
     CTF_LAPACK::cdgelsd(m, n, k, A, lda_A, B, lda_B, S, cond, rank, work, lwork, iwork, info);
 #endif
